mise en place du systemem de connexion + commentaire simple

+++ PHP/page/login.php
@@ -9,24 +9,42 @@ if(isset($_POST["usernameLogin"]) && isset($_POST["passwordLogin"])){
     $t = connect($conn, $_POST["usernameLogin"], $_POST["passwordLogin"]);
     if($t["correct"]){
         $_SESSION["user"] = $t["id"];
-        header("location: ../../html/profile.html");
+        header("Location: /mael/html/profile.html");
         exit();
     }else{
         echo "<script>alert('Ientifiant ou mot de passe incorrect');</script>";
     }
 }
 
-
 //Inscription
 if(isset($_POST["usernameSignup"]) && isset($_POST["passwordSignup"]) && isset($_POST["emailSignup"])){
-    createUser($conn, $_POST["usernameSignup"], $_POST["emailSignup"], $_POST["passwordSignup"]);
-    $t = connect($conn, $_POST["usernameSignup"], $_POST["passwordSignup"]);
-    if($t["correct"]){
-        $_SESSION["user"] = $t["id"];
-        header("Location: ../../html/profile.html");
-        exit();
-    }else{
-        echo "<script>alert('La création du compte a échoué');</script>";
+    $usernameSignup = $_POST["usernameSignup"];
+    $emailSignup = $_POST["emailSignup"];
+    
+    // Vérifier que l'email n'existe pas deja
+    $checkEmailQuery = "SELECT `id` FROM `utilisateurs` WHERE `email`='$emailSignup'";
+    $checkEmailRet = mysqli_query($conn, $checkEmailQuery);
+    $emailExists = mysqli_num_rows($checkEmailRet) > 0;
+    
+    // Vérifier que le pseudo n'existe pas deja
+    $checkPseudoQuery = "SELECT `id` FROM `utilisateurs` WHERE `pseudo`='$usernameSignup'";
+    $checkPseudoRet = mysqli_query($conn, $checkPseudoQuery);
+    $pseudoExists = mysqli_num_rows($checkPseudoRet) > 0;
+    
+    if($emailExists){
+        echo "<script>alert('Cette adresse email existe déjà');</script>";
+    } else if($pseudoExists){
+        echo "<script>alert('Ce pseudo existe déjà');</script>";
+    } else {
+        createUser($conn, $usernameSignup, $emailSignup, $_POST["passwordSignup"]);
+        $t = connect($conn, $usernameSignup, $_POST["passwordSignup"]);
+        if($t["correct"]){
+            $_SESSION["user"] = $t["id"];
+            header("Location: /mael/html/profile.html");
+            exit();
+        }else{
+            echo "<script>alert('La création du compte a échoué');</script>";
+        }
     }
 }
